# Improve RelateNG performance for A/L cases in prepared predicates