--- a/modules/core/src/test/java/org/locationtech/jts/operation/relateng/RelateNGGCTest.java
+++ b/modules/core/src/test/java/org/locationtech/jts/operation/relateng/RelateNGGCTest.java
@@ -224,12 +224,20 @@ public class RelateNGGCTest extends RelateNGTestCase {
     String a = "POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))";
     String b = "GEOMETRYCOLLECTION (POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0)), LINESTRING (0 2, 0 5))";
     checkEquals(a, b, true);
+    checkContainsWithin(a, b, true);
+    checkCoversCoveredBy(a, b, true);
+    checkContainsWithin(b, a, true);
+    checkCoversCoveredBy(b, a, true);
   }
 
   public void testPolygonContainingLineInBoundaryAndInterior() {
     String a = "POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0))";
     String b = "GEOMETRYCOLLECTION (POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0)), LINESTRING (0 2, 0 5, 5 5))";
     checkEquals(a, b, true);
+    checkContainsWithin(a, b, true);
+    checkCoversCoveredBy(a, b, true);
+    checkContainsWithin(b, a, true);
+    checkCoversCoveredBy(b, a, true);
   }