Improve the `RelateNG` self-noding check to allow A-L cases to take advantage of prepared predicates.

Previously, A/L cases were not using the prepared noder cache for predicates which require self-noding (such as `contains`, `covers`, `relate` etc).  This caused significantly poor performance for a common situation.  The self-noding check has been refined to allow A/L cases to be use the cached noder.

This makes `RelateNG` in prepared mode as fast as `PreparedGeometry` for all cases (and it handles more cases).

For example, a test in `RelateNGPolygonLinesOverlappingPerfTest` has dramatically reduce execution time (133x faster!):
```
-------  Running with A: polygon # pts = 200000   B # pts = 400  x 10000 lines
runContainsOld : 9888 ms
runContainsOldPrep : 36 ms
runContainsNG : 5906 ms

BEFORE: runContainsNGPrep : 4540 ms
NOW:    runContainsNGPrep : 34 ms
```